feat: move enclave to core
