+++ packages/@controllers/src/index.ts
@@ -1,2 +1,1 @@
 export * from "./isle";
-export * from "./enclave";
