+++ packages/@controllers/src/isle/index.ts
@@ -1,5 +1,8 @@
 import {
   type DelegatedWriteGrantSignatureRequest,
+  type EnclaveOptions,
+  type EnclaveProvider,
+  IframeEnclave,
   type IsleControllerMessage,
   type IsleMessageHandler,
   type IsleNodeMessage,
@@ -44,8 +47,6 @@ import { mainnet, sepolia } from "@wagmi/core/chains";
 import { BrowserProvider, type JsonRpcSigner } from "ethers";
 import { goTry } from "go-try";
 import invariant from "tiny-invariant";
-import { IframeEnclave } from "../enclave";
-import type { EnclaveOptions, EnclaveProvider } from "../enclave/types";
 
 /**
  * Meta information about an actor.
