+++ packages/@core/src/index.ts
@@ -6,3 +6,4 @@ export * from "./kwil-nep413-signer";
 export * from "./kwil-actions";
 export * from "./store";
 export * from "./utils";
+export * from "./enclave";
