+++ packages/idos-sdk-js/src/lib/index.ts
@@ -1,6 +1,6 @@
 export * from "./idos";
 export { default as idOSGrant } from "./grants";
-export * from "@idos-network/controllers/enclave";
+export * from "@idos-network/core/enclave";
 export * from "./utils";
 export * from "./grants";
